+++ src/main/java/kitchenpos/dao/MenuGroupDao.java
@@ -2,7 +2,7 @@ package kitchenpos.dao;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public interface MenuGroupDao {
     MenuGroup save(MenuGroup entity);
