+++ src/test/java/kitchenpos/acceptance/MenuGroupAcceptanceTest.java
@@ -7,8 +7,8 @@ import io.restassured.response.Response;
 import java.util.List;
 import kitchenpos.AcceptanceTest;
 import kitchenpos.acceptance.helper.KitchenPosBehaviors;
-import kitchenpos.domain.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
+import kitchenpos.menu.domain.MenuGroup;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
