+++ src/test/java/kitchenpos/fixture/MenuFixtureFactory.java
@@ -3,7 +3,7 @@ package kitchenpos.fixture;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
 
 public class MenuFixtureFactory {
