+++ src/test/java/kitchenpos/fixture/MenuGroupFixtureFactory.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuGroupFixtureFactory {
     private MenuGroupFixtureFactory() {
